Update GEP.gms
In this update we fixed notation, added budget constraints, fixed the production constraint, and made other small changes. We also switched from a model where the decision variable is discrete to one in which it is continuous for model development purposes. We intend to switch this back once we have the model further developed. At this point the model should compile and solve. Just to reiterate the model is a single node, two stage model with stochastic demand.