Update GEP.gms
A common technique in generation expansion planning is to model investment and capacity used as the decision variables. I thnk there is a problem with this implementation and the information we need later to calculate system reliability. I will think about this further and add more text to the document. We track the number of hours each technology is used, as well as the number of units. In the demand constraint this produces a bilinear term sum(p, x(n,p)*size(p)*y(n,p,s)) >= d(n,s) for each s. To address the bilinear term we use the McCormick envelope method. The method relaxes the non-convex problem into a convex problem to address the bilinear terms.