Update GEP.gms
Made significant progress on the GAMS and R integration. At this point GAMS conveys information on the sets to R. R takes that information and generates a new matrix and outputs it as a .csv file with the appropriate column headings. GAMS then converts the file to a .gdx file before then loading it. I am having an interesting problem that needs to be addressed. If changes are made in the R script they are not reflected in the GAMS output until the GAMS program has been run twice (i.e., first time you run the gams code old values remain). I think this has to do with the order in which I execute the commands in the GAMS code.