Update GEP.gms
Put in place the Limcol and Limrow options, which should make tracking issues easier in the future. Found a good reference to available solver and model options. Incorporated 'tryLinear' which might be of use later.  The command examines empirical NLP model to see if there are any NLP terms active. If there are none the default LP solver will be used. To activate use model_name.trylinear=1. Default value is 0. The procedure also checks to see if QCP, and DNLP models can be reduced to an LP; MIQCP and MINLP can be solved as an MIP; RMIQCP and RMINLP can be solved as an RMIP (Source: http://www.gams.com/latest/docs/userguides/userguide/_u_g__model_solve.html).